backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "asyncpg>=0.30.0",
    "fastapi[standard]>=0.115.6",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.10.4",
    "redis>=5.2.1",
    "sqlmodel>=0.0.22",
    "uvicorn>=0.34.0",
    "alembic>=1.14.0",
    "python-dotenv>=1.0.1",
    "greenlet>=3.1.1",
    "sqlalchemy[asyncio]>=2.0.36",
    "pytest-asyncio>=0.25.2",
    "httpx>=0.28.1",
    "colorama>=0.4.6",
    "markdown>=3.7",
    "marker>=2.1.3",
    "marker-pdf>=1.2.3",
    "asyncio>=3.4.3",
    "aiofiles>=24.1.0",
    "pymupdf4llm>=0.0.17",
    "itext2kg>=0.0.7",
]
